Refactored fetch-sourcemaps.js to ensure SourceMapConsumer objects can be immediately gc'ed
The way it was before, we'd load the sourcemap of each unique targetUrl into memory concurrently, and then in the end we'd actually use them for the appropriate frames. This meant that stack2source would require the combined memory footprint for all sourcemap objects to be used at the same time.

With the new way, the sourcemap object is instantiated and used in the same function and not leaked outside that context. This means Node can gc after each sourcemap is discarded (if necessary), and the total memory footprint is lowered by a lot.